+++ wwwsrc/src/services/VaultsService.js
@@ -1,6 +1,7 @@
 import { api } from './AxiosService'
 import { AppState } from '../AppState'
 import { logger } from '../utils/Logger'
+import Swal from 'sweetalert2'
 
 class VaultsService {
   // get requests
@@ -43,12 +44,46 @@ class VaultsService {
   }
 
   async deleteVault(vaultId, profileId) {
-    try {
-      await api.delete('api/vaults/' + vaultId)
-      this.getUserProfileVaults(profileId)
-    } catch (error) {
-      logger.error(error)
-    }
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
+    })
+
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then(async(result) => {
+      if (result.isConfirmed) {
+        try {
+          await api.delete('api/vaults/' + vaultId)
+          this.getUserProfileVaults(profileId)
+        } catch (error) {
+          logger.error(error)
+        }
+        swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'The vault has been deleted.',
+          'success'
+        )
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'The vault has not been deleted',
+          'error'
+        )
+      }
+    })
   }
 }
